Add indicators to location picker
