+++ apps/files/src/components/FilesLists/Indicators.vue
@@ -1,17 +1,29 @@
 <template>
   <span>
     <div v-if="displayDefaultIndicators" :class="{ 'uk-margin-xsmall-right': customIndicators }">
-      <oc-button
-        v-for="(indicator, index) in defaultIndicators"
-        :key="index"
-        class="file-row-share-indicator uk-text-middle"
-        :class="{ 'uk-margin-xsmall-left': index > 0, 'uk-invisible': !indicator.visible }"
-        :aria-label="indicator.label"
-        variation="raw"
-        @click.native.stop="indicator.handler(item, indicator.id)"
-      >
-        <oc-icon :name="indicator.icon" class="uk-text-middle" size="small" variation="system" />
-      </oc-button>
+      <template v-if="areIndicatorsClickable">
+        <oc-button
+          v-for="(indicator, index) in defaultIndicators"
+          :key="index"
+          class="file-row-share-indicator uk-text-middle"
+          :class="{ 'uk-margin-xsmall-left': index > 0, 'uk-invisible': !indicator.visible }"
+          :aria-label="indicator.label"
+          variation="raw"
+          @click.native.stop="indicator.handler(item, indicator.id)"
+        >
+          <oc-icon :name="indicator.icon" class="uk-text-middle" size="small" variation="system" />
+        </oc-button>
+      </template>
+      <template v-else>
+        <oc-icon
+          v-for="(indicator, index) in defaultIndicators"
+          :key="index"
+          :name="indicator.icon"
+          class="uk-text-middle"
+          size="small"
+          variation="system"
+        />
+      </template>
     </div>
     <template v-if="customIndicators">
       <component
@@ -38,6 +50,11 @@ export default {
     defaultIndicators: {
       type: Array,
       required: true
+    },
+    areIndicatorsClickable: {
+      type: Boolean,
+      required: false,
+      default: true
     }
   },
